+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # Konfiguracja bazy danych
 # Ustawienia polaczenia z baza danych MySQL
 spring.datasource.url=jdbc:mysql://localhost:3306/taskmanager?useSSL=false&serverTimezone=UTC&createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true
-spring.datasource.username=
+spring.datasource.username=root
 spring.datasource.password=
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
